spi_nand_flash: continue reading if ECC fails
As with erase, read operation on NAND flash may fail if a block is worn
or bad. Do not abort reading on first block, as getting even incomplete
dump is useful for data recovery.
In case of error, still return an error code to the calling process.

While at that, fix a typo in log message about ECC error.